# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/open-apis/core/rbac/system/permission": {
            "get": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "open-apis/core"
                ],
                "summary": "获取权限项列表(UserAccessToken)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页数",
                        "name": "size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "租户Pk",
                        "name": "enterprise_pk",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "操作类型",
                        "name": "operation_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "权限名称",
                        "name": "permission_name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "资源类型",
                        "name": "resource_type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "权限类型",
                        "name": "visibility",
                        "in": "query"
                    }
                ],
                "responses": {}
            },
            "put": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "open-apis/core"
                ],
                "summary": "修改权限项(UserAccessToken)",
                "parameters": [
                    {
                        "description": "请求参数体",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rbac.UpdaterPermissionReq"
                        }
                    }
                ],
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "open-apis/core"
                ],
                "summary": "新增权限项(UserAccessToken)",
                "parameters": [
                    {
                        "description": "请求参数体",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rbac.CreatePermissionReq"
                        }
                    }
                ],
                "responses": {}
            },
            "delete": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "open-apis/core"
                ],
                "summary": "删除权限项(UserAccessToken)",
                "parameters": [
                    {
                        "description": "请求参数体",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rbac.DeleteReq"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/open-apis/core/rbac/system/permission/group": {
            "get": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "open-apis/core"
                ],
                "summary": "获取权限组列表接口(UserAccessToken)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页数",
                        "name": "size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "租户Pk",
                        "name": "enterprise_pk",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "组名",
                        "name": "group_name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页数",
                        "name": "group_type",
                        "in": "query"
                    }
                ],
                "responses": {}
            },
            "put": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "open-apis/core"
                ],
                "summary": "修改权限组(UserAccessToken)",
                "parameters": [
                    {
                        "description": "请求参数体",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rbac.UpdaterPermissionGroupReq"
                        }
                    }
                ],
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "open-apis/core"
                ],
                "summary": "新增权限组(UserAccessToken)",
                "parameters": [
                    {
                        "description": "请求参数体",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rbac.CreatePermissionGroupReq"
                        }
                    }
                ],
                "responses": {}
            },
            "delete": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "open-apis/core"
                ],
                "summary": "删除权限组(UserAccessToken)",
                "parameters": [
                    {
                        "description": "请求参数体",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rbac.DeleteReq"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/open-apis/core/rbac/system/permission/group/permission": {
            "post": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "open-apis/core"
                ],
                "summary": "新增权限组关联(UserAccessToken)",
                "parameters": [
                    {
                        "description": "请求参数体",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rbac.CreatePermissionGroupPermissionReq"
                        }
                    }
                ],
                "responses": {}
            },
            "delete": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "open-apis/core"
                ],
                "summary": "删除权限组关联关系(UserAccessToken)",
                "parameters": [
                    {
                        "description": "请求参数体",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rbac.DeleteReq"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "rbac.CreatePermissionGroupPermissionReq": {
            "type": "object",
            "properties": {
                "permission_group_pk": {
                    "type": "string"
                },
                "permission_pk": {
                    "type": "string"
                }
            }
        },
        "rbac.CreatePermissionGroupReq": {
            "type": "object",
            "properties": {
                "enterprise_pk": {
                    "type": "string"
                },
                "group_name": {
                    "type": "string"
                },
                "group_type": {
                    "type": "integer"
                },
                "remark": {
                    "type": "string"
                }
            }
        },
        "rbac.CreatePermissionReq": {
            "type": "object",
            "properties": {
                "enterprise_pk": {
                    "type": "string"
                },
                "operation_type": {
                    "type": "integer"
                },
                "permission_name": {
                    "type": "string"
                },
                "resource": {
                    "type": "string"
                },
                "resource_type": {
                    "type": "integer"
                },
                "visibility": {
                    "type": "integer"
                }
            }
        },
...
        "rbac.DeleteReq": {
            "type": "object",
            "properties": {
                "pk": {
                    "type": "string"
                }
            }
        },
...
        "rbac.UpdaterPermissionGroupReq": {
            "type": "object",
            "properties": {
                "group_name": {
                    "type": "string"
                },
                "group_type": {
                    "type": "integer"
                },
                "pk": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                }
            }
        },
        "rbac.UpdaterPermissionReq": {
            "type": "object",
            "properties": {
                "operation_type": {
                    "type": "integer"
                },
                "permission_name": {
                    "type": "string"
                },
                "pk": {
                    "type": "string"
                },
                "resource": {
                    "type": "string"
                },
                "resource_type": {
                    "type": "integer"
                },
                "visibility": {
                    "type": "integer"
                }
            }
        },
...
    }
}